backend/requirements.txt
fastapi>=0.116.1
uvicorn>=0.35.0
pydantic>=2.11.7
pydantic-settings>=2.10.1
python-multipart==0.0.6
pillow==10.0.1
opencv-python==4.8.1.78
numpy==1.25.2
pandas==2.1.1
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
wordcloud==1.9.2
easyocr==1.7.0
pytesseract==0.3.10
torch==2.0.1
torchvision==0.15.2
scipy==1.11.3
python-jose==3.3.0
python-dotenv==1.0.0
aiofiles==23.2.0